requirements.txt
python-dotenv
requests>=2.21.0
requests-oauthlib>=1.2.0